squashed old and new tests
